Use size_t
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@49 49c17534-1d02-dc11-b283-0030488c597c

+++ v_network_que.c
@@ -43,7 +43,7 @@ struct VNetQueue {
 	unsigned int	slot;
 };
 
-unsigned int verse_session_get_size(void)
+size_t verse_session_get_size(void)
 {
 	const VNetQueue *queue;
